PHPStan setup

+++ src/VideoBasedMarketing/Recordings/Domain/Entity/Video.php
@@ -286,7 +286,7 @@ class Video
         nullable: true,
         options: ['unsigned' => true]
     )]
-    private ?float $assetPosterStillWebpWidth = null;
+    private ?int $assetPosterStillWebpWidth = null;
 
     public function getAssetPosterStillWebpWidth(): ?int
     {
@@ -304,7 +304,7 @@ class Video
         nullable: true,
         options: ['unsigned' => true]
     )]
-    private ?float $assetPosterStillWebpHeight = null;
+    private ?int $assetPosterStillWebpHeight = null;
 
     public function getAssetPosterStillWebpHeight(): ?int
     {
@@ -336,7 +336,7 @@ class Video
         nullable: true,
         options: ['unsigned' => true]
     )]
-    private ?float $assetPosterAnimatedWebpWidth = null;
+    private ?int $assetPosterAnimatedWebpWidth = null;
 
     public function getAssetPosterAnimatedWebpWidth(): ?int
     {
@@ -354,7 +354,7 @@ class Video
         nullable: true,
         options: ['unsigned' => true]
     )]
-    private ?float $assetPosterAnimatedWebpHeight = null;
+    private ?int $assetPosterAnimatedWebpHeight = null;
 
     public function getAssetPosterAnimatedWebpHeight(): ?int
     {
@@ -436,7 +436,7 @@ class Video
         nullable: true,
         options: ['unsigned' => true]
     )]
-    private ?float $assetOriginalWidth = null;
+    private ?int $assetOriginalWidth = null;
 
     public function getAssetOriginalWidth(): ?int
     {
@@ -454,7 +454,7 @@ class Video
         nullable: true,
         options: ['unsigned' => true]
     )]
-    private ?float $assetOriginalHeight = null;
+    private ?int $assetOriginalHeight = null;
 
     public function getAssetOriginalHeight(): ?int
     {
@@ -543,7 +543,7 @@ class Video
         nullable: true,
         options: ['unsigned' => true]
     )]
-    private ?float $assetFullWebmWidth = null;
+    private ?int $assetFullWebmWidth = null;
 
     public function getAssetFullWebmWidth(): ?int
     {
@@ -561,7 +561,7 @@ class Video
         nullable: true,
         options: ['unsigned' => true]
     )]
-    private ?float $assetFullWebmHeight = null;
+    private ?int $assetFullWebmHeight = null;
 
     public function getAssetFullWebmHeight(): ?int
     {
@@ -629,7 +629,7 @@ class Video
         nullable: true,
         options: ['unsigned' => true]
     )]
-    private ?float $assetFullMp4Width = null;
+    private ?int $assetFullMp4Width = null;
 
     public function getAssetFullMp4Width(): ?int
     {
@@ -647,7 +647,7 @@ class Video
         nullable: true,
         options: ['unsigned' => true]
     )]
-    private ?float $assetFullMp4Height = null;
+    private ?int $assetFullMp4Height = null;
 
     public function getAssetFullMp4Height(): ?int
     {
@@ -697,7 +697,7 @@ class Video
         nullable: true,
         options: ['unsigned' => true]
     )]
-    private ?float $assetPosterStillWithPlayOverlayForEmailPngWidth = null;
+    private ?int $assetPosterStillWithPlayOverlayForEmailPngWidth = null;
 
     public function getAssetPosterStillWithPlayOverlayForEmailPngWidth(): ?int
     {
@@ -715,7 +715,7 @@ class Video
         nullable: true,
         options: ['unsigned' => true]
     )]
-    private ?float $assetPosterStillWithPlayOverlayForEmailPngHeight = null;
+    private ?int $assetPosterStillWithPlayOverlayForEmailPngHeight = null;
 
     public function getAssetPosterStillWithPlayOverlayForEmailPngHeight(): ?int
     {
